requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
motor==3.7.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uvicorn==0.34.0
